pyproject.toml
[project]
name = "doc-translator-deepl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "deepl>=1.27.0",
    "python-dotenv>=1.2.1",
]
